Add files via upload
This program prompts the user to enter the number of tickets they want to purchase and then prompts them to enter the age of each visitor. The program then calculates the cost of each ticket based on the visitor's age and adds it to the total cost. If the user purchases more than 3 tickets, the program applies a 10% discount to the total cost and prints the final amount to be paid.